+++ micro-service/graphql/user/repo/repo.go
@@ -3,7 +3,7 @@ package repo
 import "github.com/flexzuu/benchmark/micro-service/graphql/user/repo/entity"
 
 type User interface {
-	Get(ID int64) (entity.User, error)
+	Get(ID int) (entity.User, error)
 	Create(Email string, Name string) (entity.User, error)
-	Delete(ID int64) error
+	Delete(ID int) error
 }
